[Synthetics] Remove FTR API specs already migrated to Scout
PR #272696 ported list_monitors, private_location_apis, synthetics_api_security
and add_monitor (permissions) to Scout, and #272740 completed synthetics_enablement,
but the FTR originals were never deleted (the deletions were dropped during the
#272740 rebase off its stacked base). Remove those redundant FTR specs and trim
`apis/synthetics/index.ts`.

- Move the shared `omitMonitorKeys`/`keyToOmitList` helpers out of the deleted
  `add_monitor.ts` into `helper/monitor_keys.ts` so the retained #258046-blocked
  specs (`add_monitor_private_location`, `sync_global_params`,
  `sync_global_params_spaces`) keep compiling.
- Drop the serverless `synthetics_enablement` suite: the Scout
  `synthetics_enablement.spec.ts` runs on serverless via
  `@local-serverless-observability_complete`.

The FTR `no permissions errors for all user` test in `synthetics_api_security.ts`
was a dead no-op (the `|| SYNTHETICS_API_URLS.SYNTHETICS_PROJECT_APIKEY` guard is
always truthy, so `assertPermissions` never ran), so no real coverage is lost.
