--- x-pack/solutions/observability/test/serverless/api_integration/test_suites/synthetics/index.ts
@@ -1,21 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import type { FtrProviderContext } from '../../ftr_provider_context';
-
-export default function ({ getService, loadTestFile }: FtrProviderContext) {
-  const esDeleteAllIndices = getService('esDeleteAllIndices');
-
-  describe('Synthetics API Tests', () => {
-    before(async () => {
-      await esDeleteAllIndices('heartbeat*');
-      await esDeleteAllIndices('synthetics*');
-    });
-
-    loadTestFile(require.resolve('./synthetics_enablement'));
-  });
-}
